+++ terratex_reallife/USER/payday/updateplaytime.lua
@@ -61,7 +61,8 @@ function controlWeber(thePlayer)
     if math.round(vioGetElementData(thePlayer, "playtime") / 60) == 25 then
         local werber = MySql.helper.getValueSync("players", "werber", { Nickname = getPlayerName(thePlayer) });
 
-        if (MySQL_DatasetExist("players", "Nickname='" .. werber .. "'")) then
+        if (MySql.helper.existSync("players", {Nickname = werber })) then
+
             if (getPlayerName(thePlayer) ~= werber) then
                 local query = "INSERT INTO gutschriften (Nickname,Geld,Grund) VALUES ('" .. werber .. "','1500','Der von dir geworbene Spieler " .. getPlayerName(thePlayer) .. " hat 25 Spielstunden erreicht!');"
                 mysql_query(handler, query)
@@ -79,7 +80,6 @@ function controlWeber(thePlayer)
         end
     end
 end
-
 
 function payday(thePlayer)
     triggerClientEvent(thePlayer, "spawnHufeisen", thePlayer)